requirements.txt
contourpy==1.1.0
cycler==0.11.0
deprecation==2.1.0
fonttools==4.42.0
importlib-resources==6.0.1
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.4
matplotlib==3.7.2
nidaqmx==0.8.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
zipp==3.16.2
